Added a function to avoid errors
It turns out that one source of errors comes from non-existing handles (i.e., due to handle name changes or account removal). I added a conditional function to look up the username and, if not found, then loop to the next one. This will also help us avoid errors from the changes in the handle set that may occur likely in real-time.